+++ src/renderer/entities/staking/api/index.ts
@@ -1,5 +1,5 @@
 export { getAvgApy, getValidatorsApy } from './apyCalculator';
 export { useStakingData } from './stakingDataService';
-export { useStakingRewards } from './stakingRewardsService';
+export { fetchStakingRewards, stakingRewardsApi } from './stakingRewardsService';
 export { validatorsService } from './validatorsService';
 export { eraService } from './eraService';
