+++ src/renderer/entities/staking/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useStakingRewards } from './useStakingRewards';
