+++ src/renderer/entities/staking/index.ts
@@ -1,5 +1,6 @@
 export * from './ui';
 export * from './api';
+export * from './hooks';
 export * from './lib/types';
 export * from './lib/constants';
 export { useValidatorsMap } from './lib/useValidatorsMap';
